requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
supabase==2.4.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.7.0
